requirements.txt
azure-identity
openai
python-dotenv
pydantic
rich
azure-ai-inference
dotenv-azd
aiohttp
openai-agents
semantic-kernel
langgraph
langchain_openai
langchain_community
azure-search-documents
ipython
azure-storage-blob
langchain-ollama
langchain_experimental
text_generation
langchain-core
langchain-tests